add injector, move injection to view, iProjectCard
