+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/inject/Injector.kt
@@ -1,0 +1,5 @@
+package org.wycliffeassociates.otter.jvm.app.ui.inject
+
+object Injector {
+    val projectDao = DaggerDBComponent().build.injectProjectDao()
+}
